requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
discord==1.0.1
discord.py==1.5.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
MarkupSafe==1.1.1
multidict==4.7.6
pycparser==2.20
python-dotenv==0.14.0
six==1.15.0
Werkzeug==1.0.1
yarl==1.6.0
